de-tangle the hafascontroller a bit more
